feat(registry-change): put requests to utils for testing
